Revert "Merge branch 'swiftui' into feature/grades-view"
This reverts commit 10ded310bfe3abc3572f134c456400542c92e525, reversing
changes made to b328053cbb347ccb4be2eeea3b698afe8430dc83.
