+++ .gitignore
@@ -62,5 +62,3 @@ Preview.html  # Preview file by deliver
 # https://github.com/johnno1962/injectionforxcode
 
 iOSInjectionProject/
-
-*.DS_Store
