--- Campus-iOS/Base/Errors/BackendError.swift
@@ -1,14 +1,0 @@
-//
-//  Errors.swift
-//  Campus-iOS
-//
-//  Created by Milen Vitanov on 15.12.21.
-//
-
-enum BackendError: Error {
-    case network(error: Error)
-    case dataSerialization(error: Error)
-    case jsonSerialization(error: Error)
-    case xmlSerialization(error: Error)
-    case objectSerialization(reason: String)
-}
