--- Campus-iOS/CalendarComponent/ViewModel/CalendarViewModel.swift
@@ -1,15 +1,0 @@
-//
-//  CalendarViewModel.swift
-//  Campus-iOS
-//
-//  Created by Milen Vitanov on 23.12.21.
-//
-
-import Foundation
-
-class CalendarViewModel: ObservableObject {
-    
-    @Published var showEventsList: Bool = false
-    
-}
-
