--- Campus-iOS/Enums/Enums.swift
@@ -1,14 +1,0 @@
-//
-//  Enums.swift
-//  Campus-iOS
-//
-//  Created by Milen Vitanov on 13.01.22.
-//
-
-import Foundation
-
-enum SheetDestination {
-    case none
-    case loginSheet(model: Model)
-    case profileSheet(model: Model)
-}
