+++ Campus-iOS/Map/Cafeteria.swift
@@ -42,7 +42,7 @@ final class Cafeteria: NSObject, Decodable, MKAnnotation {
     
 }
 
-extension Array where Element == Cafeteria {
+/*extension Array where Element == Cafeteria {
     mutating func sortByDistance(to location: CLLocation) {
         self.sort { (lhs,rhs) in
             let lhsDistance = lhs.coordinate.location.distance(from: location)
@@ -50,4 +50,4 @@ extension Array where Element == Cafeteria {
             return lhsDistance < rhsDistance
         }
     }
-}
+}*/
