+++ Campus-iOS/Map/Dish.swift
@@ -22,6 +22,8 @@ struct Price: Decodable {
 
 struct Dish: Decodable, Hashable {
 
+    
+
     /*
      {
         "name": "Pasta all'arrabiata",
