+++ Campus-iOS/ProfileComponent/ProfileModel.swift
@@ -1,0 +1,15 @@
+//
+//  ProfileModel.swift
+//  Campus-iOS
+//
+//  Created by Anton Wyrowski on 11.12.21.
+//
+
+import Foundation
+import SwiftUI
+
+class ProfileModel: ObservableObject {
+    
+    @Published var showProfile: Bool = false
+    
+}
