+++ Campus-iOS/ProfileComponent/ProfileView.swift
@@ -8,7 +8,7 @@
 import SwiftUI
 
 struct ProfileView: View {
-    @ObservedObject var model: Model
+    @EnvironmentObject var profileModel: ProfileModel
     
     var body: some View {
         
@@ -73,48 +73,19 @@ struct ProfileView: View {
                                 UIApplication.shared.open(mailToUrl, options: [:])
                         }
                     }
-                }
-                
-                Section() {
-                    HStack(alignment: .bottom) {
-                        Spacer()
-                        if model.isUserAuthenticated {
-                            Button(action: {
-                                model.logout()
-                            }) {
-                                Text("Logout").foregroundColor(.red)
-                            }
-                        } else {
-                            Button(action: {
-                                model.isLoginSheetPresented = true
-                            }) {
-                                Text("Sign in").foregroundColor(.green)
-                            }
-                        }
-                        Spacer()
+                    
+                    Button("Logout", role: .destructive) {
+                        
                     }
-                }
-            }
-            .sheet(isPresented: $model.isLoginSheetPresented) {
-                NavigationView {
-                    LoginView(model: model)
                 }
             }
             .navigationTitle("Profile")
             .navigationBarTitleDisplayMode(.inline)
             .toolbar {
-                Button(action: {model.showProfile.toggle()}) {
+                Button(action: {profileModel.showProfile.toggle()}) {
                     Text("Done").bold()
                 }
             }
         }
-    }
-}
-
-struct ProfileView_Previews: PreviewProvider {
-    static let model = MockModel()
-    
-    static var previews: some View {
-        ProfileView(model: model)
     }
 }
