+++ Campus-iOS/Styles/CustomRoundedBorderTextFieldStyle.swift
@@ -11,12 +11,9 @@ import SwiftUI
 struct CustomRoundedTextFieldStyle: TextFieldStyle {
     func _body(configuration: TextField<Self._Label>) -> some View {
         configuration
-            .padding(5)
-            .background(RoundedRectangle(cornerRadius: 5).fill(Color(UIColor.systemGray6)))
-            .overlay(
-                RoundedRectangle(cornerRadius: 5)
-                    .stroke(lineWidth: 0.5)
-                    .foregroundColor(Color(UIColor.systemGray4))
-            )
+            .padding(10)
+            .background(LinearGradient(gradient: Gradient(colors: [Color.orange, Color.orange]), startPoint: .topLeading, endPoint: .bottomTrailing))
+            .cornerRadius(20)
+            .shadow(color: .gray, radius: 10)
     }
 }
